Moving to static definition, updating test to match
