+++ .circleci/sstest.cpp
@@ -1,0 +1,6 @@
+#include "sstest.h"
+#include "storage-selector.h"
+
+BlockDevice* sstest_indirection() {
+    return storage_selector();
+}
