+++ .circleci/sstest.h
@@ -1,0 +1,3 @@
+#include "BlockDevice.h"
+
+BlockDevice* sstest_indirection();
